+++ ticketing/src/main/java/com/goti/ticketing/session/service/application/ReservationSessionService.java
@@ -3,6 +3,7 @@ package com.goti.ticketing.session.service.application;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Map;
+import java.util.Optional;
 import java.util.UUID;
 
 import org.springframework.stereotype.Service;
@@ -11,12 +12,12 @@ import org.springframework.transaction.annotation.Transactional;
 import com.goti.constants.messages.ErrorCode;
 import com.goti.exception.CustomException;
 import com.goti.global.validation.Preconditions;
+import com.goti.infra.cache.RedisCache;
 import com.goti.infra.constants.redis.RedisKey;
 import com.goti.ticketing.domain.entity.seat.SeatHoldEntity;
 import com.goti.ticketing.domain.entity.seat.SeatStatusEntity;
 import com.goti.ticketing.seat.service.domain.SeatStatusService;
 import com.goti.ticketing.session.model.ReservationSessionCache;
-import com.goti.ticketing.session.repository.ReservationSessionRedisRepository;
 import com.goti.ticketing.seat.repository.SeatHoldRepository;
 
 import lombok.RequiredArgsConstructor;
@@ -24,7 +25,9 @@ import lombok.RequiredArgsConstructor;
 @Service
 @RequiredArgsConstructor
 public class ReservationSessionService {
-	private final ReservationSessionRedisRepository reservationSessionRedisRepository;
+	private static final String KEY_DELIMITER = ":";
+
+	private final RedisCache redisCache;
 	private final SeatHoldRepository seatHoldRepository;
 	private final SeatStatusService seatStatusService;
 
@@ -41,7 +44,7 @@ public class ReservationSessionService {
 			ErrorCode.BAD_REQUEST
 		);
 
-		ReservationSessionCache reservationSession = reservationSessionRedisRepository.find(memberId, gameId)
+		ReservationSessionCache reservationSession = find(memberId, gameId)
 			.orElseGet(() -> create(memberId, gameId));
 
 		return reservationSession;
@@ -61,12 +64,12 @@ public class ReservationSessionService {
 			ErrorCode.BAD_REQUEST
 		);
 
-		ReservationSessionCache reservationSession = reservationSessionRedisRepository.find(memberId, gameId)
+		ReservationSessionCache reservationSession = find(memberId, gameId)
 			.orElseThrow(() -> new CustomException(ErrorCode.RESERVATION_SESSION_EXPIRED));
 
 		if (reservationSession.expiresAt().isBefore(LocalDateTime.now())) {
 			releaseHeldSeats(memberId, gameId);
-			reservationSessionRedisRepository.delete(memberId, gameId);
+			delete(memberId, gameId);
 			throw new CustomException(ErrorCode.RESERVATION_SESSION_EXPIRED);
 		}
 	}
@@ -76,8 +79,27 @@ public class ReservationSessionService {
 			UUID.randomUUID(),
 			LocalDateTime.now().plus(RedisKey.RESERVATION_SESSION.getTtl())
 		);
-		reservationSessionRedisRepository.save(memberId, gameId, reservationSession);
+		redisCache.set(
+			RedisKey.RESERVATION_SESSION,
+			generateKeyParam(memberId, gameId),
+			reservationSession
+		);
 		return reservationSession;
+	}
+
+	private Optional<ReservationSessionCache> find(UUID memberId, UUID gameId) {
+		return Optional.ofNullable(
+			redisCache.get(
+				RedisKey.RESERVATION_SESSION.getKey(generateKeyParam(memberId, gameId)),
+				ReservationSessionCache.class
+			)
+		);
+	}
+
+	private boolean delete(UUID memberId, UUID gameId) {
+		return redisCache.delete(
+			RedisKey.RESERVATION_SESSION.getKey(generateKeyParam(memberId, gameId))
+		);
 	}
 
 	private void releaseHeldSeats(UUID memberId, UUID gameId) {
@@ -99,5 +121,9 @@ public class ReservationSessionService {
 			}
 			seatHold.release();
 		}
+	}
+
+	private String generateKeyParam(UUID memberId, UUID gameId) {
+		return memberId + KEY_DELIMITER + gameId;
 	}
 }
